requirements-frozen.txt
-f https://packages.wdt.pdx.edu/dist/
psu.oit.arc.aol==1.5.6
amqp==5.0.5
awscli==1.19.59
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.59
botocore==1.20.59
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
cryptography==3.4.7
dateparser==1.0.0
defusedxml==0.7.1
Django==2.2.20
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-sendfile2==0.6.0
djangorestframework==3.12.4
docutils==0.15.2
idna==2.10
jmespath==0.10.0
kombu==5.0.2
oauthlib==3.1.0
oyaml==1.0
packaging==20.9
Pillow==8.1.2
psu.oit.wdt.emcee==1.1.0.b4
psycopg2==2.8.6
pwgen==0.8.2.post0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyshp==2.1.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-aws4auth==1.0.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
sentry-sdk==1.0.0
Shapely==1.7.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
terminaltables==3.1.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
